requirements.txt
# Ubuntu + NVIDIA GPU environment
# PyTorch CUDA wheel source: https://download.pytorch.org/whl/cu121
# Assumes a compatible NVIDIA driver is already installed on the target machine.

--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121

accelerate==1.10.1
bitsandbytes==0.47.0
datasets==4.0.0
peft==0.17.1
PyYAML==6.0.2
transformers==4.57.1
trl==0.25.1

PyPDF2==3.0.1
safetensors==0.6.2
scipy==1.13.1
sentencepiece==0.2.0
